rviz: progress on bug #5058 - MOVE_PLANE controls now drag based on 3D position of the grab point.
git-svn-id: https://code.ros.org/svn/ros-pkg/stacks/visualization/trunk@37820 eb33c2ac-9c88-4c90-87e0-44a10359b0c3

+++ src/rviz/default_plugin/interactive_markers/interactive_marker.cpp
@@ -434,8 +434,8 @@ bool InteractiveMarker::handleMouseEvent(ViewportMouseEvent& event, const std::s
 
     visualization_msgs::InteractiveMarkerFeedback feedback;
     feedback.event_type = (event.event.LeftDown() ?
-                           visualization_msgs::InteractiveMarkerFeedback::MOUSE_DOWN :
-                           visualization_msgs::InteractiveMarkerFeedback::MOUSE_UP);
+                           (uint8_t)visualization_msgs::InteractiveMarkerFeedback::MOUSE_DOWN :
+                           (uint8_t)visualization_msgs::InteractiveMarkerFeedback::MOUSE_UP);
                            
     feedback.control_name = control_name;
     feedback.marker_name = name_;
